Hide the lock
